Add validation to new, rename and delete graph methods. Add error code handling methods.

+++ graph_store.rb
@@ -5,6 +5,23 @@ before do
     content_type 'application/json'
 end
 
+#TODO: other message when request is not found ex: post delete
+not_found do
+  {error: "Graph not found!"}.to_json
+end
+
+error 400 do
+  {error: "Invalid request!"}.to_json
+end
+
+error 409 do
+  {error: "Graph with provided name already exists!"}.to_json
+end
+
+error 500...600 do
+  {error: "Internal server error!"}.to_json
+end
+
 get '/' do
   "List api info/documentation"
   #TODO: OR list documentation
@@ -15,24 +32,41 @@ get '/list' do #or /graphs
   "list"
 end
 
-#TODO: add new graph
-#after that add nodes and edges
 post '/new' do
   graph_name = params[:name]
-  File.write(graph_name, params.to_json)
+
+  halt 400 if graph_name == nil
+
+  if File.exist?(graph_name)
+    halt 409
+  else
+    File.write(graph_name, params.to_json)
+  end
 end
 
-#TODO: rename existing graph
 put '/rename' do
   old_name = params[:old_name]
   new_name = params[:new_name]
-  File.rename(old_name, new_name)
+
+  halt 400 if old_name == nil || new_name == nil
+
+  if File.exist?(old_name)
+    File.rename(old_name, new_name)
+  else
+    halt 404
+  end
 end
 
-#TODO: delete graph
 delete '/delete' do
   graph_name = params[:name]
-  File.delete(graph_name)
+
+  halt 400 if graph_name == nil
+
+  if File.exist?(graph_name)
+    File.delete(graph_name)
+  else
+    halt 404
+  end
 end
 
 #TODO:search graph
